Add scoring to game
Using the SFML Text object and loading in a custom font.

Currently the score is updated when a ball is spawned into the
screen. However, it might make more sense to have it increment
when a ball leaves the court instead. Have not implemented a
way to check if ball leaves the screen.