Reduce bug where player continues moving
Player continues to move even if no keys are pressed.
Moved player controls back into event loop to increase chances
of KeyReleased event being detected.

However, this does not completely solve the issue as unable to
directly query if keyboard button is being pressed or not on mac.
(Issue encountered on a previous commit for player movement)