Change start game keys
Previously used enter/return key to begin the game. This method
creates some difficulty in starting because the player needs to
switch rapidly from pressing enter to using the arrow keys.

Now use arrow keys to start a game. Makes it easier to get used to
moving around, especially when game immediately throws a ball at
the player.

Another method would probably be to include some sort of delay when
starting a game but this shall do for now.