Change how player interacts with screen
Instead of storing screen resolution, use an IntRect which defines
the bounds of the screen with top, left, width and height.
Spawn player using this information passed in as the 'arena' to be
remembered by the player.

Makes it slightly clearer when checking if the player moves out of
screen in update function.