Add dodgeball sprite
Now uses a sprite to draw dodgeballs. The sprite uses a texture
that is 20x20 pixels, which is the size of the smallest dodgeball.
As the size of a dodgeball grows, the sprite is scaled accordingly.

A ball now has both a Sprite for visuals, as well as a CircleShape
object for doing the collision detection.