Make scoring more efficient
Score was being updated every single frame. This is unnecessary
as the score does not change that often, but only when dodgeballs
leave the screen.

Moved the stringstream code into said section which handles
dodgeballs missing. Now score is modified only every few
seconds, as opposed to hundred/thousands of times a second.