Change sprite colour on hit
Modifies the sprite colour to change to blue on hitting the
player. SFML requires a white sprite in order to change
sprite colours, because Sprite.setColor() acts as a filter
to which hue is allowed to be seen. (i.e., a red sprite will
not be able to have blue colour)

Colour change might help in identifying which ball has
the player hit (although this may not be the best way or
even necessary, and may be removed later).