Adjust dodgeball speed and difficulty
Changed dodgeball initial speed to 400px per second, slower than
the player which is 500px per second. Also increased the interval for
increasing the number of dodgeballs from 10 to 15 seconds.

Hope to smoothen the difficulty curve to allow player to remain in the
game for longer and reach a higher number of dodgeballs more
consistently. Needs further tweaking.