Refactor enemy update code
Code was checking if enemy is not active, calculate score, spawn
new enemy, which is not the most executed case. Main case is
rather, if enemy active, move enemy.

This refactor inverts the starting if else condition from !enemy.isActive()
to instead just enemy.isActive, which makes it easier to read.