Add sprite for player
Change player from a white square to an animated human. The sprite
sheet consists of 10 frames, and is 50x500 pixels in size. Each frame
has a length of 50 pixels.

Every 0.1 seconds, the texture coordinates cycle to the next frame in
the sprite sheet using the setTextureRect() function, cycling back to 0
after the last frame at (450, 0) is reached. The time accumulated is
stored in the variable m_TimeSinceLastSpriteUpdate.

When the player moves left, the sprite is flipped using setScale(). But,
using this function causes the sprite to be drawn to the left from -50 to
0 instead of from 0 to 50 on the x-axis. In order to fix this, a call to
setOrigin() is made in the flipSprite() function, which moves the sprite
back to the correct position.