Add run animation
Switches the animation to running when the player is moving.
Done by changing the texture used by calling setTexture on a pre-
loaded texture.

In the animate() function, first checks if the player is moving, and
then switches the texture accordingly.

The flipSprite() call was relocated from being called when a player
moved left or right, to being called only after the animation step.
This was because the player was not being flipped after a change
in texture, despite player already moving in the opposite direction.

Finally, sprite sheet cycling was not cycling due to run and idle sheets
having different number of frames, and checking if the last frame was
reached using left == 350 was not working. Changed the condition to
use >= instead to fix.