Change player center
Instead of a square drawn from the top left from (0, 0) to the
bottom right at (50, 50), set the center of the player directly in
the middle of the square instead.

This changes the code for preventing the player from moving
out of the screen. Rather than checking using the player's top
left corner, have to measure from the player's center and half
the length of the player.